requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
altair==6.0.0
annotated-types==0.7.0
asttokens==3.0.1
attrs==25.4.0
backcall==0.2.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
cachetools==7.0.5
catboost==1.2.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.5
curl_cffi==0.13.0
cycler==0.12.1
db-dtypes==1.5.0
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.2.1
fastjsonschema==2.21.2
finmind==1.9.6
flake8==7.3.0
Flask==3.1.3
fonttools==4.62.1
frozendict==2.4.7
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.41
google-api-core==2.30.0
google-auth==2.49.1
google-auth-oauthlib==1.3.0
google-cloud-bigquery==3.40.1
google-cloud-core==2.5.0
google-cloud-storage==3.10.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
graphviz==0.21
greenlet==3.3.2
grpcio==1.78.0
grpcio-status==1.78.0
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
ipykernel==7.2.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
kiwisolver==1.5.0
lightgbm==4.6.0
loguru==0.7.3
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mccabe==0.7.0
mistune==3.2.0
multidict==6.7.1
multitasking==0.0.12
narwhals==2.18.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.26.4
nvidia-nccl-cu12==2.29.7
oauthlib==3.3.1
optuna==4.7.0
packaging==26.0
pandas==2.3.3
pandas-gbq==0.34.0
pandocfilters==1.5.1
parso==0.8.6
patsy==1.0.2
peewee==4.0.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==12.1.1
pipreqs==0.5.0
platformdirs==4.9.2
plotly==6.6.0
pluggy==1.6.0
prettytable==3.17.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.6
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydata-google-auth==1.9.1
pydeck==0.9.1
pyecharts==2.1.0
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytoolconfig==1.3.1
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rope==1.14.0
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
soupsieve==2.8.3
SQLAlchemy==2.0.48
stack-data==0.6.3
statsmodels==0.14.0
streamlit==1.55.0
streamlit-option-menu==0.4.0
ta==0.5.25
tabulate==0.10.0
tenacity==9.1.4
terminaltables==3.1.10
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
wcwidth==0.6.0
webencodings==0.5.1
websockets==16.0
Werkzeug==3.1.6
xgboost==3.2.0
yarg==0.1.9
yarl==1.23.0
yfinance==1.2.0
zipp==3.23.0
